Update documentation to reflect context-based validation API
- Rewrite CLAUDE.md to document input-first parameter style
- Update README.md with context-based validation examples
- Replace ObjectSchema examples with schema() function usage
- Document new validation entry points: tryValidate/validate
- Update Ktor integration to use Validated interface
- Remove references to removed APIs (IdentityValidator, ElvisValidator)
- Simplify API documentation focusing on Validation extension functions
- Add factory validation and conditional validation documentation
