+++ kova-core/src/main/kotlin/org/komapper/extension/validator/LiteralValidator.kt
@@ -5,14 +5,12 @@ package org.komapper.extension.validator
  *
  * Example:
  * ```kotlin
- * val validator = Kova.string().literal("admin")
- * validator.validate("admin") // Success
- * validator.validate("user")  // Failure
+ * tryValidate { literal("admin", "admin") } // Success
+ * tryValidate { literal("user", "admin") }  // Failure
  * ```
  *
  * @param value The expected value
  * @param message Custom error message provider
- * @return A new validator that accepts only the specified value
  */
 @IgnorableReturnValue
 fun <S> Validation.literal(
@@ -26,14 +24,13 @@ fun <S> Validation.literal(
  *
  * Example:
  * ```kotlin
- * val validator = Kova.string().literal(listOf("admin", "user", "guest"))
- * validator.validate("admin") // Success
- * validator.validate("other") // Failure
+ * val allowed = listOf("admin", "user", "guest")
+ * tryValidate { literal("admin", allowed) } // Success
+ * tryValidate { literal("other", allowed) } // Failure
  * ```
  *
  * @param values The list of acceptable values
  * @param message Custom error message provider
- * @return A new validator that accepts only values from the list
  */
 @IgnorableReturnValue
 fun <S> Validation.literal(
