+++ kova-core/src/test/kotlin/org/komapper/extension/validator/CharSequenceValidatorTest.kt
@@ -5,56 +5,17 @@ import io.kotest.core.spec.style.FunSpec
 class CharSequenceValidatorTest :
     FunSpec({
 
-        context("plus") {
-            fun Validation.validate(string: String) {
-                max(string, 2)
-                max(string, 3)
-            }
-
-            test("success") {
-                val result = tryValidate { validate("1") }
-                result.shouldBeSuccess()
-            }
-
-            test("failure") {
-                val result = tryValidate { validate("1234") }
-                result.shouldBeFailure()
-                result.messages.size shouldBe 2
-                result.messages[0].constraintId shouldBe "kova.charSequence.max"
-                result.messages[1].constraintId shouldBe "kova.charSequence.max"
-            }
-        }
-
-        context("and") {
-            fun Validation.validate(string: String) {
-                max(string, 2)
-                max(string, 3)
-            }
-
-            test("success") {
-                val result = tryValidate { validate("1") }
-                result.shouldBeSuccess()
-            }
-
-            test("failure") {
-                val result = tryValidate { validate("1234") }
-                result.shouldBeFailure()
-                result.messages.size shouldBe 2
-                result.messages[0].constraintId shouldBe "kova.charSequence.max"
-                result.messages[1].constraintId shouldBe "kova.charSequence.max"
-            }
-        }
-
-        context("then") {
+        context("with conversion") {
             fun Validation.validate(string: String) =
                 string.trim().let { trimmed ->
                     length(trimmed, 3)
-                    trimmed.toUppercase()
+                    trimmed.uppercase()
                 }
 
             test("success") {
                 val result = tryValidate { validate(" abc ") }
                 result.shouldBeSuccess()
+                result.value shouldBe "ABC"
             }
 
             test("failure") {
@@ -277,61 +238,6 @@ class CharSequenceValidatorTest :
                 val result = tryValidate { max1("12") }
                 result.shouldBeFailure()
                 result.messages.single().constraintId shouldBe "kova.charSequence.max"
-            }
-        }
-
-        context("trim with constraints") {
-            test("success when trimmed value meets constraint") {
-                val result = tryValidate { min("  hello  ".trim(), 3) }
-                result.shouldBeSuccess()
-            }
-
-            test("failure when trimmed value violates constraint") {
-                val result = tryValidate { min("  hi  ".trim(), 3) }
-                result.shouldBeFailure()
-                result.messages.single().constraintId shouldBe "kova.charSequence.min"
-            }
-
-            test("failure when whitespace only becomes empty after trim") {
-                val result = tryValidate { min("   ".trim(), 3) }
-                result.shouldBeFailure()
-                result.messages.single().constraintId shouldBe "kova.charSequence.min"
-            }
-        }
-
-        context("toUpperCase with constraints") {
-            test("success when transformed value meets constraint") {
-                val result = tryValidate { min("hello".toUppercase(), 3) }
-                result.shouldBeSuccess()
-            }
-
-            test("failure when transformed value violates constraint") {
-                val result = tryValidate { min("hi".toUppercase(), 3) }
-                result.shouldBeFailure()
-                result.messages.single().constraintId shouldBe "kova.charSequence.min"
-            }
-
-            test("success when combining toUpperCase with startsWith") {
-                val result = tryValidate { startsWith("hello".toUppercase(), "H") }
-                result.shouldBeSuccess()
-            }
-        }
-
-        context("toLowerCase with constraints") {
-            test("success when transformed value meets constraint") {
-                val result = tryValidate { min("HELLO".toLowercase(), 3) }
-                result.shouldBeSuccess()
-            }
-
-            test("failure when transformed value violates constraint") {
-                val result = tryValidate { min("HI".toLowercase(), 3) }
-                result.shouldBeFailure()
-                result.messages.single().constraintId shouldBe "kova.charSequence.min"
-            }
-
-            test("success when combining toLowerCase with startsWith") {
-                val result = tryValidate { startsWith("HELLO".toLowercase(), "h") }
-                result.shouldBeSuccess()
             }
         }
     })
