+++ kova-core/src/test/kotlin/org/komapper/extension/validator/LiteralValidatorTest.kt
@@ -2,7 +2,7 @@ package org.komapper.extension.validator
 
 import io.kotest.core.spec.style.FunSpec
 
-class IdentityValidatorTest :
+class LiteralValidatorTest :
     FunSpec({
 
         context("literal") {
@@ -74,67 +74,6 @@ class IdentityValidatorTest :
                     result.messages.size shouldBe 1
                     result.messages[0].constraintId shouldBe "kova.literal.list"
                 }
-            }
-        }
-
-        context("onlyIf") {
-            fun Validation.validate(i: Int) {
-                if (i % 2 == 0) min(i, 3)
-            }
-            test("success when condition not met") {
-                val result = tryValidate { validate(1) }
-                result.shouldBeSuccess()
-            }
-
-            test("failure when condition met") {
-                val result = tryValidate { validate(2) }
-                result.shouldBeFailure()
-                result.messages.size shouldBe 1
-                result.messages[0].constraintId shouldBe "kova.comparable.min"
-            }
-
-            context("with plus") {
-                fun Validation.validateAndMin1(i: Int) {
-                    if (i % 2 == 0) min(i, 3)
-                    min(i, 1)
-                }
-
-                test("success") {
-                    val result = tryValidate { validateAndMin1(1) }
-                    result.shouldBeSuccess()
-                }
-
-                test("failure") {
-                    val result = tryValidate { validateAndMin1(0) }
-                    result.shouldBeFailure()
-                    result.messages.size shouldBe 2
-                    result.messages[0].constraintId shouldBe "kova.comparable.min"
-                    result.messages[1].constraintId shouldBe "kova.comparable.min"
-                }
-            }
-        }
-
-        context("constrain") {
-            @IgnorableReturnValue
-            fun Validation.validate(i: Int) = i.constrain("even") { satisfies(it % 2 == 0) { text("input must be even") } }
-
-            test("failure") {
-                val result = tryValidate { validate(1) }
-                result.shouldBeFailure()
-                result.messages.size shouldBe 1
-                result.messages[0].text shouldBe "input must be even"
-            }
-        }
-
-        context("constrain with extension function") {
-            @IgnorableReturnValue
-            fun Validation.even(i: Int) = i.constrain("even") { satisfies(it % 2 == 0) { text("input must be even") } }
-
-            test("failure") {
-                val result = tryValidate { even(1) }
-                result.shouldBeFailure()
-                result.messages.size shouldBe 1
-                result.messages[0].text shouldBe "input must be even"
             }
         }
     })
