+++ kova-core/src/test/kotlin/org/komapper/extension/validator/StringValidatorTest.kt
@@ -5,48 +5,6 @@ import io.kotest.core.spec.style.FunSpec
 class StringValidatorTest :
     FunSpec({
 
-        context("or") {
-            fun Validation.validate(string: String): String {
-                val _ = or { isInt(string) } orElse { literal(string, "zero") }
-                return string.toUppercase()
-            }
-
-            test("success with int value") {
-                val result = tryValidate { validate("1") }
-                result.shouldBeSuccess()
-                result.value shouldBe "1"
-            }
-
-            test("success with literal value") {
-                val result = tryValidate { validate("zero") }
-                result.shouldBeSuccess()
-                result.value shouldBe "ZERO"
-            }
-
-            test("failure") {
-                val result = tryValidate { validate("abc") }
-                result.shouldBeFailure()
-                result.messages.size shouldBe 1
-                result.messages[0].constraintId shouldBe "kova.or"
-            }
-        }
-
-        context("constrain") {
-            @IgnorableReturnValue
-            fun Validation.validate(string: String) = string.constrain("test") { satisfies(it == "OK") { text("Constraint failed") } }
-
-            test("success") {
-                val result = tryValidate { validate("OK") }
-                result.shouldBeSuccess()
-            }
-
-            test("failure") {
-                val result = tryValidate { validate("NG") }
-                result.shouldBeFailure()
-                result.messages.single().text shouldBe "Constraint failed"
-            }
-        }
-
         context("notBlank with message") {
             test("success") {
                 val result = tryValidate { notBlank("ab") { text("Must not be blank") } }
@@ -357,108 +315,6 @@ class StringValidatorTest :
                 val result = tryValidate { "0".stringBools() }
                 result.shouldBeSuccess()
                 result.value shouldBe false
-            }
-        }
-
-        context("trim") {
-            test("success when trimming leading whitespace") {
-                val result = tryValidate { "  hello".trim() }
-                result.shouldBeSuccess()
-                result.value shouldBe "hello"
-            }
-
-            test("success when trimming trailing whitespace") {
-                val result = tryValidate { "hello  ".trim() }
-                result.shouldBeSuccess()
-                result.value shouldBe "hello"
-            }
-
-            test("success when trimming both sides") {
-                val result = tryValidate { "  hello  ".trim() }
-                result.shouldBeSuccess()
-                result.value shouldBe "hello"
-            }
-
-            test("success with no whitespace to trim") {
-                val result = tryValidate { "hello".trim() }
-                result.shouldBeSuccess()
-                result.value shouldBe "hello"
-            }
-
-            test("success with empty string") {
-                val result = tryValidate { "".trim() }
-                result.shouldBeSuccess()
-                result.value shouldBe ""
-            }
-
-            test("success with only whitespace") {
-                val result = tryValidate { "   ".trim() }
-                result.shouldBeSuccess()
-                result.value shouldBe ""
-            }
-        }
-
-        context("toUpperCase") {
-            test("success when converting lowercase to uppercase") {
-                val result = tryValidate { "hello".toUppercase() }
-                result.shouldBeSuccess()
-                result.value shouldBe "HELLO"
-            }
-
-            test("success when converting mixed case to uppercase") {
-                val result = tryValidate { "HeLLo".toUppercase() }
-                result.shouldBeSuccess()
-                result.value shouldBe "HELLO"
-            }
-
-            test("success with already uppercase") {
-                val result = tryValidate { "HELLO".toUppercase() }
-                result.shouldBeSuccess()
-                result.value shouldBe "HELLO"
-            }
-
-            test("success with empty string") {
-                val result = tryValidate { "".toUppercase() }
-                result.shouldBeSuccess()
-                result.value shouldBe ""
-            }
-
-            test("success with numbers and symbols") {
-                val result = tryValidate { "hello123!@#".toUppercase() }
-                result.shouldBeSuccess()
-                result.value shouldBe "HELLO123!@#"
-            }
-        }
-
-        context("toLowerCase") {
-            test("success when converting uppercase to lowercase") {
-                val result = tryValidate { "HELLO".toLowercase() }
-                result.shouldBeSuccess()
-                result.value shouldBe "hello"
-            }
-
-            test("success when converting mixed case to lowercase") {
-                val result = tryValidate { "HeLLo".toLowercase() }
-                result.shouldBeSuccess()
-                result.value shouldBe "hello"
-            }
-
-            test("success with already lowercase") {
-                val result = tryValidate { "hello".toLowercase() }
-                result.shouldBeSuccess()
-                result.value shouldBe "hello"
-            }
-
-            test("success with empty string") {
-                val result = tryValidate { "".toLowercase() }
-                result.shouldBeSuccess()
-                result.value shouldBe ""
-            }
-
-            test("success with numbers and symbols") {
-                val result = tryValidate { "HELLO123!@#".toLowercase() }
-                result.shouldBeSuccess()
-                result.value shouldBe "hello123!@#"
             }
         }
